requirements.txt
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.0
flower==1.2.0
idna==3.4
ipython==8.11.0
jedi==0.18.2
lxml==4.9.2
matplotlib-inline==0.1.6
numpy==1.24.2
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
snscrape==0.6.1.20230314
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.25.0
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6